Moved markdown parser from comments to helper and added it to submission and challenge descriptions
